[#5] move : 보드 패키지 구조 변경
- 도메인 패키지 아래에 하위 패키지 작성
- 패키지 변경으로 인한 파일 이동 및 패키지 경로 수정
