+++ src/main/java/com/sparta/tentrillion/stat/Stat.java
@@ -1,6 +1,6 @@
 package com.sparta.tentrillion.stat;
 
-import com.sparta.tentrillion.board.Board;
+import com.sparta.tentrillion.board.entity.Board;
 import com.sparta.tentrillion.card.Card;
 import com.sparta.tentrillion.global.TimeStamp;
 import jakarta.persistence.*;
